[Modify] Map 페이지 코드 분리 (useMap hook 생성)
